Move environment extraction from extras to pyproject.rs
